requirements.txt
asgiref==3.5.0
certifi==2021.10.8
Django==4.0.3
djangorestframework==3.13.1
mysqlclient==2.1.0
pytz==2022.1
sentry-sdk==1.5.8
sqlparse==0.4.2
urllib3==1.26.9
